ultimos cambios

+++ src/main/java/com/example/utnfinaljava/dtos/ProductSupplierDto.java
@@ -12,5 +12,5 @@ public class ProductSupplierDto {
     private Integer amount;
     private String supplierName;
     private List<PriceDto> prices;
-    private Long validityPrice;
+    private Float validityPrice;
 }
